+++ src/main/java/com/sanluan/einvoice/utils/StringUtils.java
@@ -1,0 +1,8 @@
+package com.sanluan.einvoice.utils;
+
+public class StringUtils {
+
+    public static String replace(String str) {
+        return str.replaceAll(" ", "").replaceAll("　", "").replaceAll("：", ":").replaceAll(" ", "");
+    }
+}
